+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.2.0] - 2018-07-26
+
+#### Added 
+* database migration command
+* you can now also just pass a list of field names to exporter `$fields` argument
+
+#### Changed
+* ExportTypeItemInterface::prepareItemFields arguments order
+
 ## [0.1.0] - 2018-07-25
 
 Initial release.